feat: Add post detail page and route
Add a new route and template for displaying the details of a post. This includes creating a new template file called "post_detail.html" and updating the "routes.py" file to include the new route for accessing the post detail page. This will allow users to view the full content of a post and interact with it.